Created Login Modal
Created the login modal that allows us to access linkedin's api when a user logs in. Only certain web domains are authorized to use our key to access the API so the application will only work either on davidcorey.org, coreycurse.org, or localhost so you can run it on your own machine if you launch a server.